Returns 1 when benchmarking throws an exception (#136)

+++ llm_bench/python/llm_run_bloomz_560m_on_linux.sh
@@ -19,11 +19,22 @@ dst_dir="./ov_models/bloomz-560m"
 convert_model="python ./llm_bench/python/convert.py --model_id ${original_dir} --output_dir ${dst_dir} --precision FP16 --stateful"
 echo ${convert_model}
 eval ${convert_model}
+ret=$?
+if [ ${ret} -ne 0 ]; then
+    echo "Convert bloomz-560m failed, ret=${ret}"
+    exit ${ret}
+fi
 wait
 
-bemchmarking="python ./llm_bench/python/benchmark.py -m ${dst_dir}/pytorch/dldt/FP16/ -d cpu -n 1"
-echo ${bemchmarking}
-eval ${bemchmarking}
+benchmarking="python ./llm_bench/python/benchmark.py -m ${dst_dir}/pytorch/dldt/FP16/ -d cpu -n 1"
+echo ${benchmarking}
+eval ${benchmarking}
+ret=$?
 
 rm -rf ${original_dir}
 rm -rf ${dst_dir}
+
+if [ ${ret} -ne 0 ]; then
+    echo "Benchmarking failed, ret=${ret}"
+    exit ${ret}
+fi
