+++ llm_bench/python/llm_run_stable_diffusion_on_linux.sh
@@ -16,14 +16,25 @@ wait
 original_dir="tiny-sd"
 dst_dir="./ov_models/tiny-sd"
 
-convert_model="python ./llm_bench/python/convert.py --model_id ${original_dir} --output_dir ${dst_dir} --precision FP16 --stateful"
+convert_model="python ./llm_bench/python/convert.py --model_id ${original_dir} --output_dir ${dst_dir} --precision FP16"
 echo ${convert_model}
 eval ${convert_model}
+ret=$?
+if [ ${ret} -ne 0 ]; then
+    echo "Convert tiny-sd failed, ret=${ret}"
+    exit ${ret}
+fi
 wait
 
-bemchmarking="python ./llm_bench/python/benchmark.py -m ${dst_dir}/pytorch/dldt/FP16/ -pf ./llm_bench/python/prompts/stable-diffusion.jsonl -d cpu -n 1"
-echo ${bemchmarking}
-eval ${bemchmarking}
+benchmarking="python ./llm_bench/python/benchmark.py -m ${dst_dir}/pytorch/dldt/FP16/ -pf ./llm_bench/python/prompts/stable-diffusion.jsonl -d cpu -n 1"
+echo ${benchmarking}
+eval ${benchmarking}
+ret=$?
 
 rm -rf ${original_dir}
 rm -rf ${dst_dir}
+
+if [ ${ret} -ne 0 ]; then
+    echo "Benchmarking failed, ret=${ret}"
+    exit ${ret}
+fi
